+++ src/main/java/backend/AI/AIMemory.java
@@ -1,4 +1,4 @@
-package metrics;
+package AI;
 
 import data.Constants;