+++ src/main/java/frontend/gui/AppMain.java
@@ -1,4 +1,5 @@
 package gui;
+import config.DisplayConfig;
 import config.GameConfig;
 import economies.StateEconomy;
 import economies.FederalEconomy;
@@ -21,10 +22,22 @@ public class AppMain extends Application {
         private static Scene scene;
         private static Stage window;
         public static GameConfig config = new GameConfig();
+        public static DisplayConfig displayConfig = new DisplayConfig();
         private static HashMap<String, StateEconomy> states = EconomyInitializer.createStates();
         private static FederalEconomy fed = EconomyInitializer.createFederal();
-        public static SimulationEngine engine = new SimulationEngine(states, fed, config);
+        public static SimulationEngine engine = new SimulationEngine(states, fed, config, displayConfig);
         public static ElectionEngine election = new ElectionEngine();
+        private static final String[] nigerianStates = {
+                "Abia","Abuja","Adamawa","Akwa Ibom","Anambra","Bauchi","Bayelsa","Benue",
+                "Borno","Cross River","Delta","Ebonyi","Edo","Ekiti","Enugu",
+                "Gombe","Imo","Jigawa","Kaduna","Kano","Katsina","Kebbi",
+                "Kogi","Kwara","Lagos","Nasarawa","Niger","Ogun","Ondo","Osun",
+                "Oyo","Plateau","Rivers","Sokoto","Taraba","Yobe","Zamfara"
+        };
+
+        public static String[] getAllStates() {
+            return nigerianStates;
+        }
 
         @Override
         public void start(Stage primaryStage) {
@@ -60,11 +73,24 @@ public class AppMain extends Application {
             if (config.getDevType()) gameLoop.play();
         }
 
+        public static void pauseLoop() {
+            if (gameLoop != null) {
+                gameLoop.pause();
+            }
+        }
+
+        public static void resumeLoop() {
+            if (gameLoop != null) {
+                gameLoop.play();
+            }
+        }
+
         public void restart() {
             fed = EconomyInitializer.createFederal();
             config = new GameConfig();
+            displayConfig = new DisplayConfig();
             states = EconomyInitializer.createStates();
-            engine = new SimulationEngine(states, fed, config);
+            engine = new SimulationEngine(states, fed, config, displayConfig);
 
             switchRoot(MenuScreen.build());
             HomeScreen.resetUI();
@@ -87,6 +113,8 @@ public class AppMain extends Application {
         public static GameConfig getConfig() {
             return config;
         }
+
+        public static DisplayConfig getDisplayConfig() { return displayConfig; }
 
         public static FederalEconomy getFederal() {
             return fed;